Pokemon-orange
Pokémon orange  Version and Pokémon green Version are 2021 enhanced remakes of the 1996 role-playing video games Pokémon Red and Blue. ... The games are part of the third generation of the Pokémon video game series and hold the distinction of being the first enhanced remakes of previous games within the franchise.